JUGBodensee/TweetwallFX#15 renamed tweet-spi-twitter to tweet-impl-twitter4j in accordance to comment (package name changed as well, with the correct name and correct package declaration)
